🔨 Refactor: 플레이어 스크립트 하드코딩 제거

+++ Project  G/Assets/1.Script/InGameManager/ScoreManager.cs
@@ -59,7 +59,7 @@ public class ScoreManager : Singleton<ScoreManager>
         isReadyToCount = true;
     }
 
-    public void AddScore(int score) 
+    public void AddScore(float score) 
     {
         Debug.Log($"점수증가 : {currScore} + {score} = {currScore + score}");
         currScore += score;
